Permission

+++ commands/clear.js
@@ -2,10 +2,17 @@ module.exports = {
   name: 'clear',
   description: 'Untuk membersihkan chat',
   execute(message, args) {
-    if (!args[1]) {
-      message.reply('Masukkan jumlah pesan yang ingin dihapus!');
+    const role = message.member.roles.cache.find(
+      (role) => role.id === '856900957757767710'
+    );
+    if (role) {
+      if (!args[1]) {
+        message.reply('Masukkan jumlah pesan yang ingin dihapus!');
+      } else {
+        message.channel.bulkDelete(args[1]);
+      }
     } else {
-      message.channel.bulkDelete(args[1]);
+      message.reply('Mohon maaf kamu bukan satpam di server ini!');
     }
   },
 };
